Add files via upload
My corpus is made up of songs found in playlists found on Spotify defined as 'autumn' playlists. this graph depicts the valence on the x-axis and the danceability on the y-axis. It shows that the happier the song, the higher the danceability, which was to be expected. However, is also a big cluster of songs that are classified as having a low valence that still score somewhat high on the danceability scale. Therefore, one of the assumptions that can be made is that some 'autumn' songs can be classified as negative sounding with still some level of danceability.